job_runner/Raw_Gent/sub_agents/bug_fix_workflow/bug_fix_sub_agents/fix_code/fix_code_agent.py
from google.adk.agents import LlmAgent
from .fix_code_prompt import Fix_Code_Prompt

Fix_Code_Agent = LlmAgent(
    name="fix_code",
    model="gemini-3.1-flash-lite-preview",
    description="",
    instruction=Fix_Code_Prompt
)