job_runner/requirements.txt
google-cloud-run
google-cloud-storage
google-cloud-aiplatform
google-adk
requests
httpx
google-genai
google-cloud-logging
redis